Feat: Magazine 정보를 가져오는 UseCase 생성
