+++ core-model/src/main/java/com/hmoa/core_model/data/MagazineContentItem.kt
@@ -1,0 +1,7 @@
+package com.hmoa.core_model.data
+
+data class MagazineContentItem(
+    var header : String? = null,
+    var content : String? = null,
+    var image : String? = null
+)
